+++ .covignore
@@ -1,3 +1,3 @@
 ._mock\.go
 .test_fixture\.go
-./testing/.
+test/.
