+++ cmd/file-store-svc/main_test.go
@@ -8,7 +8,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/kinneko-de/restaurant-file-store-svc/internal/testing/envvar"
+	"github.com/kinneko-de/restaurant-file-store-svc/test/testing/envvar"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
