+++ test/testing/io/write_closer_mock.go
